requirements.txt
# Core Framework
scrapy==2.11.0
selenium==4.15.0
beautifulsoup4==4.12.2
lxml==5.0.0
requests==2.31.0
requests-html==0.10.0

# Async & Task Queue
celery==5.3.4
flower==2.0.1
redis==5.0.1

# Anti-Bot & Automation
undetected-chromedriver==3.5.4
fake-useragent==1.4.0
playwright==1.40.0
2captcha-python==1.2.1

# Data Quality & Validation
pydantic==2.5.2
pandas==2.1.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
phonenumbers==8.13.27

# Databases
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pymongo==4.6.1
opensearch-py==2.4.2

# Message Queue
pika==1.3.2
kombu==5.3.4

# Retry & Resilience
tenacity==8.2.3
backoff==2.2.1

# HTTP & Network
httpx==0.25.2
aiohttp==3.9.1

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk==1.39.1

# Dashboard
streamlit==1.28.2
plotly==5.18.0
altair==5.2.0

# API (Optional)
fastapi==0.108.0
uvicorn[standard]==0.25.0

# Data Processing
datasketch==1.6.4
python-dateutil==2.8.2

# Utilities
python-dotenv==1.0.0
click==8.1.7
pyyaml==6.0.1
